Port phoneME to the PS4 as PSNokia
Runs J2ME (MIDP) games natively on a jailbroken PS4, built with the
OpenOrbis toolchain; each game is packaged as its own PS4 app.

- 64-bit CLDC-HI VM: the VM keeps its 32-bit heap layout with 4-byte
  narrow references (narrow<T>), ROM pointers emitted as 32-bit values and
  ROMStructs fields as __ptr32, and all VM memory kept below 2GB by a
  dlmalloc-based low heap (ps4/common).
- PS4 platform layers for the VM (cldc/src/vm/os/ps4), PCSL and MIDP:
  SDL2 video scaled to the TV, DS4 input with a Nokia-style soft key
  layout, per-game save directories and screen sizes, and fixes for
  OpenOrbis header mismatches (struct stat, open flags, sigaction).
- LP64 fixes in shared MIDP code (CRCs, KNI handle macros).
- MIDI music through a new software synthesizer (midi_synth.c).
- M3G loader and renderer fixes (vertex array layout, textures, node
  transforms) and Nokia DirectGraphics counter-clockwise rotations.
- ps4/: build scripts, compiler wrappers, packaging and tools.

Includes the Vita work since the VitoKia release; the Vita README moves
to README-VITA.md.

+++ phoneme_feature/pcsl/memory/heap/pcsl_chunkheap.c
@@ -76,7 +76,7 @@ static ChunkInfo *get_chunk_info(void* chunk_ptr) {
  * to or greater than <code>ptr</code> 
  */
 void* align_size_up(void *ptr, unsigned int alignment) {
-    return (void*)(((unsigned int)ptr + alignment - 1) & ~(alignment - 1));
+    return (void*)(((size_t)ptr + alignment - 1) & ~(size_t)(alignment - 1));
 }
 
 /*
